Fix bug causing validators to not be set with def>
I noticed that my validators weren't getting set. I looked into it a bit
and discovered a macro-expansion bug. Previously the macro expansion
looked like this:

(let*
  [bpath__27248__auto__
   (str "/" (string/replace (str *ns*) #"\." "/") "/" 'validated-value)
   z__27249__auto__
   (apply
     >-
     bpath__27248__auto__
     [2 4 6]
     (:validator #(every? even? %)))]
  (when {}
    (znode/add-descendant (.znode z__27249__auto__) "/.metadata" {}))
  (def validated-value z__27249__auto__))

with the problematic line being

(:validator #(every? even? %)))

This expression was being evaluated as the keyword function and
returning nil.

I changed the options to a vec to avoid evaluation and passing the
arguments through to the >- function