Fix bug causing validators to not be set with def>
Previously the macro expansion looked like this:

(let*
  [bpath__27248__auto__
   (str "/" (string/replace (str *ns*) #"\." "/") "/" 'validated-value)
   z__27249__auto__
   (apply
     >-
     bpath__27248__auto__
     [2 4 6]
     (:validator #(every? even? %)))]
  (when {}
    (znode/add-descendant (.znode z__27249__auto__) "/.metadata" {}))
  (def validated-value z__27249__auto__))

with the problematic line being

(:validator #(every? even? %)))

This expression was being evaluated as the keyword function and
returning nil.

Now we expand the options directly into the `>-` function arguments.
This addresses the bug identified in PR GH-2.

Thanks for Adam Frey for identifying this bug and outlining the solution.