Defend against data-less nodes in node/get-data response handling
Bizzarely, znodes can be created not only with an empty byte array,
but also with nil (are these somehow serialized the same way over the
wire?) and the result is a node with no data to be
deserialized/decoded.  This actually occurs with the *old* zookeeper
client by Liebke that is used during testing.